fix: anti-hallucination + robustness hardening across MCP data paths
Five-agent read-only audit (stats, EDF/binary, provenance, services, security)
followed by fixes in TDD batches. Full suite: 387 tests green.

Anti-hallucination (no fabricated clinical values reach the LLM):
- ChannelPercentiles empty -> NaN (was 0); propagates as "no data", not a measured 0
- OscarChannelUnitNormalizer: blank-unit flow keeps unknown unit, not a fabricated L/min
- OscarEventSummaryBuilder: surface EVE<->summary count delta + reason, not a bare boolean
- UnifiedNightAnalysisService: gate provenance/data_sources on actually-loaded stats,
  not file presence, so the dormant OSCAR-EDF branch cannot claim available=true
- OscarRepository: histogram bucket Math.toIntExact (degrade, not silent int wrap)

Binary parse robustness (live SleepHQ S3 download path, not OSCAR/SQLite):
- EdfAnnotationParser: a malformed TAL skips one event instead of aborting the night
- EdfBinarySupport.scale: reject inverted/non-finite ranges (no sign-flipped values)
- ViatomSessionParser: O2 duration from decoded samples, not pre-trim record count
- EdfParser: long byte offsets + capacity guard; BinaryDownloadSupport carries fileId

Waveform/comparison/tools:
- WaveformWindowPlanner: "peak leak" excludes min-kind notable moments
- ComparisonService: one bad day -> skipped row, cannot sink the whole comparison
- WaveformTools/Planner: bound startMinute/maxMinutes before *60 (int overflow)

Security:
- McpResponses: stop echoing unexpected-exception raw messages to the client
- Documented anon-flag (loopback only) + SleepHQ password-grant contract

Reuse: consolidate (ObjectNode) JsonApi.parse casts into JsonApi.parseObject()
(also fixes 2 previously-unguarded sites).

+++ src/main/java/com/adriangarett/sleephqmcp/oscar/OscarRepository.java
@@ -324,7 +324,9 @@ public class OscarRepository {
 
                     TreeMap<Integer, Long> buckets = merged.computeIfAbsent(scRow.code(), k -> new TreeMap<>());
                     for (long[] h : hist) {
-                        buckets.merge((int) h[0], h[1], Long::sum);
+                        // Fail loud rather than silently wrap a >2^31 histogram code into a negative bucket
+                        // (which would corrupt every percentile). Real OSCAR channel codes fit in int.
+                        buckets.merge(Math.toIntExact(h[0]), h[1], Long::sum);
                     }
                 }
             }
